SCRUM-42 [Group Wallet] Settlement Recording Handler

+++ backend/core-api/internal/domain/dto/transaction-dto.go
@@ -8,7 +8,7 @@ import (
 type TransactionDto struct {
 	ID         int64                 `json:"id"`
 	Title      *string               `json:"title"`
-	Price      *float32              `json:"price"`
+	Price      *float64              `json:"price"`
 	DateMade   *time.Time            `json:"date_made"`
 	CategoryId *int64                `json:"category_id"`
 	Type       *enum.TransactionType `json:"type"`
