+++ backend/core-api/internal/domain/model/transaction.go
@@ -8,7 +8,7 @@ import (
 type Transaction struct {
 	ID         int64                `json:"id"`
 	Title      string               `json:"title"`
-	Price      float32              `json:"price"`
+	Price      float64              `json:"price"`
 	DateMade   time.Time            `json:"date_made"`
 	OwnerId    string               `json:"owner_id"`
 	CategoryId *int64               `json:"category_id"`
